Heatmap of H matrix for Joint NMF Visualization

This function is designed to allow you to visualize the weights of the H matrix from the output of the RcppML package. By specifying group labels in the function you can easily see how your sample weights relate to variables of interest

Usage

H_heatmap(
  model,
  group_labels = NULL,
  group_colors = NULL,
  show_rownames = FALSE,
  cluster_cols = FALSE,
  cluster_rows = TRUE,
  clustering_method = "ward.D2",
  main = "Heatmap of Samples, H matrix",
  cutree_rows = 4,
  ...
)

Arguments

model

NMF object generated by RcppML

group_labels

Sample annotations for the heatmap. Input can be a vector or list of vectors properly aligned to samples in your H matrix.

group_colors

If you want to manually specify colors for the row annotations. Input can be a vector or a list of vectors

show_rownames

Show feature names in heatmap, default to FALSE

cluster_cols

Clustering of NMF components, default to FALSE

cluster_rows

Clustering of Samples, default to TRUE

clustering_method

Heatmap clustering method, default to ward.D2, options from pheatmap

main

Heatmap title

cutree_rows

Specify number of sample clusters in the heatmap

...

Arguments can be passed to the pheatmap function

Value

Heatmap of H (Sample) weights matrix

Details

The primary input is an nmf object from the RcppML::nmf function. You can also specify labels of interest, which should be in a single vector or list of vectors and aligned to the H matrix. If you'd like to specify colors for the labels, you can also do so in a single vector or list. Refer to the pheatmap function for other parameters to further customize the heatmap.