Function to merge metadata from two datasets and extract label of interest

For downstream joint NMF analysis, you need to align labels of interest to the samples in your input data, but often the metadata is stored in multiple dataframes corresponding to the different data types. This provides an easy workaround to extract a metadata label of interest and align it to the combined input data file in one line.

Usage

merge_metadata(metadata1, metadata2, id_col, data_matrix, metadata_columns)

Arguments

metadata1

First metadata table

metadata2

Second metadata table

id_col

Name of the column in metadata that contains sample IDs ("Sample_ID"). Must be the same for both dataframes

data_matrix

NMF input data or combined dataframe, with column names corresponding to your samples

metadata_columns

Metadata column(s) of interest

Value

A vector of your aligned group label of choice, or a list of vectors if there is more than one group

Details

This function requires two dataframes containing sample metadata. Note that your "group of interest" and sample "ID column" MUST have the same name in both dataframes if you want group labels from both dataframes. Labels only present in one data type are acceptable and NA values will be used for any missing information.